Created experiments submodule and moved EpochExp to its own file.
