+++ examples/quad_3D.py
@@ -11,7 +11,7 @@ import numpy as np
 import pybullet as p
 import matplotlib.pyplot as plt
 
-from safe_control_gym.experiment import Experiment
+from safe_control_gym.experiments.experiment import Experiment
 from safe_control_gym.utils.configuration import ConfigFactory
 from safe_control_gym.utils.registration import make
