+++ examples/tracking.py
@@ -9,7 +9,7 @@ Run as:
 
 from functools import partial
 
-from safe_control_gym.experiment import Experiment
+from safe_control_gym.experiments.experiment import Experiment
 from safe_control_gym.utils.configuration import ConfigFactory
 from safe_control_gym.utils.registration import make
