+++ safe_control_gym/__init__.py
@@ -2,3 +2,4 @@ from safe_control_gym.controllers import *
 from safe_control_gym.safety_filters import *
 from safe_control_gym.envs import *
 from safe_control_gym.envs.env_wrappers import *
+from safe_control_gym.experiments import *
